Refacto SelectAccountAndCurrency in a single comp
